+++ frontend/src/pages/admin/submissions/Submissions.tsx
@@ -1,6 +1,6 @@
 import { Submission } from 'abacus'
 import React, { ChangeEvent, useState, useEffect, useMemo, useContext } from 'react'
-import { Button, Checkbox, Label, Menu, MenuItemProps, Table } from 'semantic-ui-react'
+import { Button, Checkbox, Grid, Label, Menu, MenuItemProps, Table } from 'semantic-ui-react'
 import Moment from 'react-moment'
 import { Link } from 'react-router-dom'
 import config from 'environment'
@@ -82,22 +82,33 @@ const Submissions = (): JSX.Element => {
     )
 
   const deleteSelected = async () => {
-    setDeleting(true)
-    const submissionsToDelete = submissions
-      .filter((submission) => submission.checked && (!submission.released || showReleased))
-      .map((submission) => submission.sid)
-    const response = await fetch(`${config.API_URL}/submissions`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${localStorage.accessToken}`
-      },
-      body: JSON.stringify({ sid: submissionsToDelete })
-    })
-    if (response.ok) {
-      loadSubmissions()
+    if (window.confirm('are you sure you want to delete these submissions?')) {
+      //if the user selects ok, then the code below runs, otherwise nothing occurs
+      setDeleting(true)
+      const submissionsToDelete = submissions
+        .filter((submission) => submission.checked && (!submission.released || showReleased))
+        .map((submission) => submission.sid)
+      const response = await fetch(`${config.API_URL}/submissions`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${localStorage.accessToken}`
+        },
+        body: JSON.stringify({ sid: submissionsToDelete })
+      })
+      if (response.ok) {
+        loadSubmissions()
+        //tells the toast container below to display a message saying 'Deleted selected submissions'
+        const id = submissionsToDelete.join()
+        window.sendNotification({
+          id,
+          type: 'success',
+          header: 'Success!',
+          content: 'We deleted the submissions you selected!'
+        })
+      }
+      setDeleting(false)
     }
-    setDeleting(false)
   }
 
   const handleItemClick = (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>, { name }: MenuItemProps) =>
@@ -114,7 +125,7 @@ const Submissions = (): JSX.Element => {
   if (isLoading) return <PageLoading />
 
   return (
-    <>
+    <Grid>
       <Helmet>
         <title>Abacus | Admin Submissions</title>
       </Helmet>
@@ -216,7 +227,7 @@ const Submissions = (): JSX.Element => {
           </Table.Body>
         </Table>
       </Block>
-    </>
+    </Grid>
   )
 }
